Minor correction for endingIndex
Removed unnecessary static_cast<double>

+++ SpeechMarker/renderarea.cpp
@@ -137,7 +137,7 @@ void RenderArea::drawSamples(QPainter &painter)
         }
         // ------------------------------------------------
         int startingIndex = qAbs(static_cast<double>(leftVisibleBorder) / samplesPerPixelValues[vectExtremaIndex]);
-        int endingIndex = qAbs(static_cast<double>(rightVisibleBorder) / static_cast<double>(samplesPerPixelValues[vectExtremaIndex]));
+        int endingIndex = qAbs(static_cast<double>(rightVisibleBorder) / samplesPerPixelValues[vectExtremaIndex]);
         QVector<QPoint> pointsToDraw((endingIndex - startingIndex + 1)*2);    // in every pixel we draw maximal and minimal values
         for(int i=startingIndex; i<=endingIndex; i++){
             // Minimal value for i-th window:
